Create controllers and controller bugfix

+++ src/main/java/problemsolved/filingsystem/controllers/HolidayRequestController.java
@@ -17,6 +17,7 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import problemsolved.filingsystem.entities.HolidayRequest;
 import problemsolved.filingsystem.entities.User;
@@ -34,20 +35,19 @@ public class HolidayRequestController {
     @GetMapping("")
     public ResponseEntity<List<HolidayRequest>> getAll() {
         Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-        List<String> roles = auth.getAuthorities().stream().map(GrantedAuthority::getAuthority).collect(Collectors.toList());
-        String username = auth.getName();
-        Optional<User> user = userRepository.findByUsername(username);
+        Optional<User> user = userRepository.findByUsername(auth.getName());
         if (user.isPresent()) {
             return ResponseEntity.ok(user.get().getHolidayRequests());
-            
         }
         return ResponseEntity.notFound().build();
     }
     
     @GetMapping("/{id}")
     public ResponseEntity<HolidayRequest> get(@PathVariable Integer id) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        Optional<User> user = userRepository.findByUsername(auth.getName());
         Optional<HolidayRequest> oRequest = holidayRepository.findById(id);
-        if (oRequest.isPresent()) {
+        if (oRequest.isPresent() && user.isPresent() && user.get().getHolidayRequests().contains(oRequest.get())) {
             return ResponseEntity.ok(oRequest.get());
         } else {
             return ResponseEntity.notFound().build();
@@ -55,15 +55,36 @@ public class HolidayRequestController {
     }
     
     @PostMapping("")
-    public ResponseEntity<HolidayRequest> insert(@RequestBody HolidayRequest issue) {
-        return ResponseEntity.ok(holidayRepository.save(issue));
+    public ResponseEntity<HolidayRequest> insert(@RequestBody HolidayRequest request) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        Optional<User> user = userRepository.findByUsername(auth.getName());
+        if(user.isPresent()) {
+            request.setStatus(HolidayRequest.Status.UNSEEN);
+            request.setUser(user.get());
+            return ResponseEntity.ok(holidayRepository.save(request));
+        }
+        return ResponseEntity.notFound().build();
+    }
+    
+    @PutMapping("/{id}")
+    public ResponseEntity<HolidayRequest> update(@PathVariable Integer id, @RequestBody HolidayRequest request) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        Optional<User> user = userRepository.findByUsername(auth.getName());
+        Optional<HolidayRequest> oRequest = holidayRepository.findById(id);
+        if (oRequest.isPresent() && user.isPresent() && user.get().getHolidayRequests().contains(oRequest.get())) {
+            request.setId(id);
+            return ResponseEntity.ok(holidayRepository.save(request));
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
     
     @DeleteMapping("/{id}")
-    //@Secured({ "ADMIN" })
     public ResponseEntity<HolidayRequest> delete(@PathVariable Integer id) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        Optional<User> user = userRepository.findByUsername(auth.getName());
         Optional<HolidayRequest> oRequest = holidayRepository.findById(id);
-        if (oRequest.isPresent()) {
+        if (oRequest.isPresent() && user.isPresent() && user.get().getHolidayRequests().contains(oRequest.get())) {
             holidayRepository.deleteById(id);
             return ResponseEntity.ok().build();
         } else {
